added gpu filters

+++ src/io/ImageHandler.h
@@ -20,10 +20,10 @@ class ImageHandler : public Singleton<ImageHandler>
     friend class Singleton<ImageHandler>;
 public:
     GLuint getImageHandleFromFilePath(std::string filePath);
-    bool loadImage(std::string filePath, std::vector<uchar>* data, int& width, int& height, int& channelNumber);
-    bool saveImage(std::string filePath, uchar* data, int width, int height, int channelNumber);
-    void replaceImageWithoutChangeOnGpu(GLuint* imageHandleGpu, uchar* data, int width, int height, int channelNumber);
-    void replaceImageWithChangeOnGpu(GLuint* imageHandleGpu, uchar* data, int width, int height, int channelNumber);
+    bool loadImage(std::string filePath, std::vector<float>* data, int& width, int& height, int& channelNumber);
+    bool saveImage(std::string filePath, float* data, int width, int height, int channelNumber);
+    void replaceImageWithoutChangeOnGpu(GLuint* imageHandleGpu, float* data, int width, int height, int channelNumber);
+    void replaceImageWithChangeOnGpu(GLuint* imageHandleGpu, float* data, int width, int height, int channelNumber);
 private:
     ImageHandler() {};
     ~ImageHandler() {};
